+++ sidApi/hw/KpixHistogram.h
@@ -12,6 +12,7 @@
 //-----------------------------------------------------------------------------
 // Modification history :
 // 10/15/2008: created
+// 03/03/2009: changed container values to unsigned int
 //-----------------------------------------------------------------------------
 #ifndef __KPIX_HISTOGRAM_H__
 #define __KPIX_HISTOGRAM_H__
@@ -26,7 +27,7 @@ using namespace std;
 class KpixHistogram {
 
       // Histogram contents
-      unsigned short *data;
+      unsigned int *data;
 
       // Number of entries in histogram
       unsigned int entries;
@@ -44,22 +45,22 @@ class KpixHistogram {
       ~KpixHistogram ( );
 
       // Add an entry
-      void fill(unsigned short value);
+      void fill(unsigned int value);
 
       // Get Number Of Entries
       unsigned int binCount();
 
       // Get Min Value
-      unsigned short minValue();
+      unsigned int minValue();
 
       // Get Max Value
-      unsigned short maxValue();
+      unsigned int maxValue();
 
       // Get Bin Value
-      unsigned short value(unsigned int bin);
+      unsigned int value(unsigned int bin);
 
       // Get Bin Count
-      unsigned short count(unsigned int bin);
+      unsigned int count(unsigned int bin);
 };
 
 #endif
